organized in messages component folder; small component styling updates
