+++ src/components/message/MessagesList.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import { Image, FlatList } from 'react-native';
-import UserRow from './UserRow';
 import PropTypes from 'prop-types';
+
+import MessageRow from '../message/MessageRow';
+
 
 const imageStyle = {
     height: 40,
@@ -13,12 +15,12 @@ const imageStyle = {
 class MessagesList extends React.Component {
     renderItem = ({item}) => {
         return (
-        <UserRow
+        <MessageRow
             senderName={item.senderName}
             msgPreviewText={item.msgPreviewText}
             // TODO: figure out way to use image links and not react nodes
             // reason is b/c require takes in a string literal and is not dynamic
-            image={<Image style={imageStyle} source={require('../assets/fade.jpg')} />}
+            image={<Image style={imageStyle} source={require('../../assets/fade.jpg')} />}
             timeStamp={item.timeStamp} 
         />
     )};
